updated AdminDao, moved Admin class to /model
