+++ src/main/java/pl/coderslab/model/Admin.java
@@ -1,4 +1,4 @@
-package pl.coderslab.dao;
+package pl.coderslab.model;
 
 import lombok.*;
